Вынес мапперы из слайса
