+++ packages/client/src/entities/forum/index.ts
@@ -1,5 +1,7 @@
-export { createCommentForTopic } from './lib/createCommentForTopic'
-export { getCommentsByTopic } from './lib/getCommentsByTopic'
+export {
+  createCommentForTopic,
+  getCommentsByTopic,
+} from './lib/commentForTopic'
 export { ForumCard } from './ui/ForumCard'
 export type { ForumCardProps } from './ui/ForumCard'
 export { CommentCard } from './ui/CommentCard/CommentCard'
