+++ packages/client/src/entities/forum/lib/commentForTopic.ts
@@ -1,4 +1,4 @@
-import { createCommentForTopicRequest } from '../api'
+import { createCommentForTopicRequest, getCommentsByTopicRequest } from '../api'
 import { ForumTopicComment } from '../model/types'
 import { commentDtoToView } from './mappers'
 
@@ -17,5 +17,18 @@ export async function createCommentForTopic(
   } catch (err) {
     console.error('Ошибка создания комментария', err)
     return null
+  }
+}
+
+export async function getCommentsByTopic(
+  topicId: number
+): Promise<ForumTopicComment[]> {
+  try {
+    const response = (await getCommentsByTopicRequest(topicId)) ?? []
+
+    return response.map(dto => commentDtoToView(dto))
+  } catch (err) {
+    console.error('Ошибка получения списка топиков')
+    return []
   }
 }
