+++ packages/client/src/entities/forum/lib/mappers.ts
@@ -1,5 +1,5 @@
-import { CommentDto } from '../api/types'
-import { ForumTopicComment } from '../model/types'
+import { CommentDto, TopicDto } from '../api/types'
+import { ForumTopic, ForumTopicComment } from '../model/types'
 
 export const commentDtoToView = (comment: CommentDto): ForumTopicComment => ({
   id: comment.id,
@@ -8,4 +8,10 @@ export const commentDtoToView = (comment: CommentDto): ForumTopicComment => ({
   lastName: String(comment.userId),
   avatarUrl: '',
   message: comment.content,
+})
+
+export const mapTopicDtoToForumTopic = (topic: TopicDto): ForumTopic => ({
+  id: String(topic.id),
+  title: topic.title,
+  text: topic.content,
 })
